fix: resolve test compilation errors
- Add missing assertFalse import in ChatSessionTest
- Add missing type cast for convertValueToTargetType assertion
- Add runTest wrapper for tests using Flow.first() suspend calls

+++ Android/src/app/src/test/java/com/server/edge/gallery/data/ConvertValueToTargetTypeTest.kt
@@ -96,7 +96,7 @@ class ConvertValueToTargetTypeTest {
 
     @Test
     fun `convert to DOUBLE from Float`() {
-        assertEquals(1.5, convertValueToTargetType(1.5f, ValueType.DOUBLE), 1e-10)
+        assertEquals(1.5, convertValueToTargetType(1.5f, ValueType.DOUBLE) as Double, 1e-10)
     }
 
     @Test
